pyproject.toml
[tool.poetry]
name = "repl_python3_gdm"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
jishaku = "^1.20.0"
django-buttons = "^0.9.2"
buttons = "^0.1.9"
flask = "^2.0.0"
chatbot-py = "^2.74"
wikipedia = "^1.4.0"
selenium = "^3.141.0"
selenium_wrapper = "^0.1"
rsap = "^2.4.2"
droblox = "^0.0.3"
aiofiles = "^0.7.0"
regex = "^2021.7.6"
aiofile = "^3.6.0"
"discord.py" = "^1.7.3"
youtube_dl = "^2021.6.6"
PyNaCl = "^1.4.0"
pyfiglet = "^0.8.post1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
